Fix ADE-74 platform service regressions
