+++ apps/desktop/src/main/services/appControl/appControlService.ts
@@ -30,6 +30,16 @@ import type {
 } from "../../../shared/types";
 import type { Logger } from "../logging/logger";
 import type { createPtyService } from "../pty/ptyService";
+import { imageDimensions } from "../shared/imageDimensions";
+import {
+  commandForwardsAppControlDebug,
+  commandLooksLikeDirectElectronLaunch,
+  commandLooksLikePackageScriptLaunch,
+  insertDebugFlagsIntoDirectElectronCommand,
+  rewritePackageScriptElectronLaunch,
+  shellQuote,
+  unquoteShellValue,
+} from "./appControlLaunchCommand";
 
 const CDP_POLL_MS = 500;
 const CDP_HEALTH_POLL_MS = 2_000;
@@ -157,86 +167,6 @@ function roundFrame(frame: AppControlFrame): AppControlFrame {
     width: round(frame.width),
     height: round(frame.height),
   };
-}
-
-function shellQuote(value: string): string {
-  if (/^[A-Za-z0-9_/:=.,@%+-]+$/.test(value)) return value;
-  return `'${value.replace(/'/g, "'\\''")}'`;
-}
-
-function commandForwardsAppControlDebug(command: string): boolean {
-  return /\{ADE_APP_CONTROL_DEBUG_FLAGS\}|\bADE_APP_CONTROL_(?:DEBUG_FLAGS|CDP_PORT|REMOTE_DEBUGGING_PORT)\b|--remote-debugging-port\b/.test(command);
-}
-
-function commandLooksLikePackageScriptLaunch(command: string): boolean {
-  return /(?:^|[;&|]\s*)(?:[A-Za-z_][A-Za-z0-9_]*=(?:"[^"]*"|'[^']*'|[^\s;&|]+)\s+)*(?:(?:npm|pnpm|yarn|bun)\s+(?:run\s+)?[A-Za-z0-9:_./-]+)\s*$/.test(command.trim());
-}
-
-function commandLooksLikeDirectElectronLaunch(command: string): boolean {
-  return /(?:^|[;&|]\s*)(?:[A-Za-z_][A-Za-z0-9_]*=(?:"[^"]*"|'[^']*'|[^\s;&|]+)\s+)*(?:npx\s+)?electron(?:\s+[^;&|]*)?\s*$/.test(command.trim());
-}
-
-function unquoteShellValue(value: string): string {
-  const trimmed = value.trim();
-  if (
-    (trimmed.startsWith("'") && trimmed.endsWith("'"))
-    || (trimmed.startsWith("\"") && trimmed.endsWith("\""))
-  ) {
-    return trimmed.slice(1, -1);
-  }
-  return trimmed;
-}
-
-function insertDebugFlagsIntoDirectElectronCommand(command: string, debugFlags: string[]): string {
-  const flags = debugFlags.map(shellQuote).join(" ");
-  return command.replace(
-    /((?:^|[;&|]\s*)(?:[A-Za-z_][A-Za-z0-9_]*=(?:"[^"]*"|'[^']*'|[^\s;&|]+)\s+)*(?:npx\s+)?electron)(?=\s|$)/,
-    `$1 ${flags}`,
-  );
-}
-
-function prependEnvToShellSegments(script: string, envPrefix: string): string {
-  const trimmedEnv = envPrefix.trim();
-  if (!trimmedEnv) return script;
-  return script
-    .split(/(\s*&&\s*)/)
-    .map((segment) => {
-      if (/^\s*&&\s*$/.test(segment)) return segment;
-      const trimmed = segment.trim();
-      return trimmed ? `${trimmedEnv} ${trimmed}` : segment;
-    })
-    .join("");
-}
-
-function rewritePackageScriptElectronLaunch(command: string, debugFlags: string[], fallbackCwd: string): string | null {
-  const match = command.trim().match(/^(?<prefix>.*?)(?<env>(?:[A-Za-z_][A-Za-z0-9_]*=(?:"[^"]*"|'[^']*'|[^\s;&|]+)\s+)*)(?<manager>npm|pnpm|yarn|bun)\s+(?:run\s+)?(?<script>[A-Za-z0-9:_./-]+)\s*$/);
-  const groups = match?.groups;
-  if (!groups) return null;
-  const prefix = groups.prefix ?? "";
-  const envPrefix = groups.env?.trim() ?? "";
-  const scriptName = groups.script;
-  if (!scriptName) return null;
-
-  let packageDir = fallbackCwd;
-  const cdMatches = Array.from(prefix.matchAll(/(?:^|[;&|]\s*)cd\s+((?:"[^"]+"|'[^']+'|[^\s;&|]+))\s*&&/g));
-  const lastCd = cdMatches.at(-1);
-  if (lastCd?.[1]) packageDir = path.resolve(fallbackCwd, unquoteShellValue(lastCd[1]));
-
-  try {
-    const packageJson = JSON.parse(fs.readFileSync(path.join(packageDir, "package.json"), "utf8")) as {
-      scripts?: Record<string, unknown>;
-    };
-    const script = packageJson.scripts?.[scriptName];
-    if (typeof script !== "string") return null;
-    if (commandForwardsAppControlDebug(script) || !/\belectron(?:\s|$)/.test(script)) return null;
-    const rewrittenScript = insertDebugFlagsIntoDirectElectronCommand(script, debugFlags);
-    if (rewrittenScript === script) return null;
-    const packageBinPath = path.join(packageDir, "node_modules", ".bin");
-    const expandedEnvPrefix = [`PATH=${shellQuote(packageBinPath)}:$PATH`, envPrefix].filter(Boolean).join(" ");
-    return `${prefix}${prependEnvToShellSegments(rewrittenScript, expandedEnvPrefix)}`;
-  } catch {
-    return null;
-  }
 }
 
 function delay(ms: number): Promise<void> {
@@ -447,48 +377,6 @@ class CdpClient {
       }
     });
   }
-}
-
-function pngDimensions(buffer: Buffer): { width: number; height: number } | null {
-  if (buffer.length < 24) return null;
-  if (buffer.toString("ascii", 1, 4) !== "PNG") return null;
-  return {
-    width: buffer.readUInt32BE(16),
-    height: buffer.readUInt32BE(20),
-  };
-}
-
-function jpegDimensions(buffer: Buffer): { width: number; height: number } | null {
-  if (buffer.length < 4 || buffer[0] !== 0xff || buffer[1] !== 0xd8) return null;
-  let offset = 2;
-  while (offset + 9 < buffer.length) {
-    if (buffer[offset] !== 0xff) {
-      offset += 1;
-      continue;
-    }
-    const marker = buffer[offset + 1];
-    offset += 2;
-    if (marker === 0xd9 || marker === 0xda) break;
-    if (offset + 2 > buffer.length) return null;
-    const segmentLength = buffer.readUInt16BE(offset);
-    if (segmentLength < 2 || offset + segmentLength > buffer.length) return null;
-    const isStartOfFrame =
-      marker >= 0xc0
-      && marker <= 0xcf
-      && ![0xc4, 0xc8, 0xcc].includes(marker);
-    if (isStartOfFrame && segmentLength >= 7) {
-      return {
-        height: buffer.readUInt16BE(offset + 3),
-        width: buffer.readUInt16BE(offset + 5),
-      };
-    }
-    offset += segmentLength;
-  }
-  return null;
-}
-
-function imageDimensions(buffer: Buffer): { width: number; height: number } | null {
-  return pngDimensions(buffer) ?? jpegDimensions(buffer);
 }
 
 function cdpDomSnapshotScript(maxElements: number): string {
